+++ CHANGELOG.md
@@ -8,6 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for specifying patterns with the `-e` / `--regexp` flag (can be used multiple times).
+- Support for reading patterns from a file with the `-f` / `--file` flag.
+
+### Changed
+
+- Made the positional pattern argument optional by adding `nargs="?"`
+
 ## [0.3.9] - 2025-10-25
 
 ### Added